some bikeshedding/documentation to improve the demo's Kafka messages
